Updated color descr's thanks to CX's help, also removed GM_addStyle polyfill
CX's issue with GM_addStyle not working turned out to be because he was using GreaseMonkey, which deprecated the GM_ methods in favor of await'ed GM. methods. His problem was fixed by switching to a different userscript manager that has better backwards compatibility. This won't be a problem for new users because the instructions recommend tampermonkey, which still supports the GM_ methods.